requirements.txt
# DLP Gateway Dependencies
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
redis[asyncio]>=5.0
rapidfuzz>=3.0
pydantic>=2.0
httpx>=0.27
python-dotenv>=1.0
pyyaml>=6.0
gliner>=0.1.0
